% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_exampleData}
\alias{df_exampleData}
\title{Example data for SPC charting}
\format{
A data frame with 41 rows and 2 variables:
\describe{
\item{Date.Time}{price, in US dollars}
\item{Value}{Values which are to be plotted on the SPC chart}
}
}
\usage{
df_exampleData
}
\description{
A dataset containing some random data useful for testing ggRFLspc
}
\keyword{datasets}
