% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRFLspc.R
\name{ggRFLspc}
\alias{ggRFLspc}
\title{create an SPC chart}
\usage{
rfl_spc_plot(data, title='Data about Measure 1', subtitle='Collected at Hospital xxx 2018/19')
}
\arguments{
\item{data}{data.frame consisting of only a date/time column in POSIXct format and a numerical column}

\item{workingFile}{Path to working file}

\item{pointSize}{Size fo the alerts and geom_points}

\item{...}{Variables sent to labs() for example title = "" and subtitle = "".}

\item{templateFile}{Path to SPC Reporting Template file}
}
\value{
a ggplot SPC chart
}
\description{
Create SPC chart from start to end
}
