% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_calculate_wb.R
\name{rfl_calculate_wb}
\alias{rfl_calculate_wb}
\title{Open, evaluate and save the SPC working file
This function calcuates the values required for the SPC by opening, evaluating and saving the the SPC template in excel}
\usage{
rfl_calculate_wb('fileToCalculate.xlsx')
}
\arguments{
\item{workingFile}{Path to working file wiht data loaded into it from rfl_load_spc_model_input}
}
\value{
location of the workingFile with the SPC calculations complete
}
\description{
Open, evaluate and save the SPC working file
This function calcuates the values required for the SPC by opening, evaluating and saving the the SPC template in excel
}
