% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_get_spc_model_output.R
\name{rfl_get_spc_model_output}
\alias{rfl_get_spc_model_output}
\title{Get data out of the SPC tempalate
This function loads the data back into R from the evaluated Excel SPC template}
\usage{
rfl_get_spc_model_output(workingFile)
}
\arguments{
\item{workingFile}{Path to working file with data loaded into it and evaluated by rfl_calculate_wb}
}
\value{
a data.frame including the original data and the calculated variables for the SPC chart
}
\description{
Get data out of the SPC tempalate
This function loads the data back into R from the evaluated Excel SPC template
}
