% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_load_spc_model_input.R
\name{rfl_load_spc_model_input}
\alias{rfl_load_spc_model_input}
\title{Input data into SPC template}
\usage{
rfl_load_spc_model_input(df_exampleData, tempfile())
}
\arguments{
\item{data}{data.frame consisting of only a date/time column in POSIXct format and a numerical column}

\item{workingFile}{Path to working file}

\item{templateFile}{Path to SPC Reporting Template file}
}
\value{
location of the workingFile with the data loaded into it
}
\description{
This function loads data into the SPC template ready for calcualtion
}
