% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_spc_plot.R
\name{rfl_spc_plot}
\alias{rfl_spc_plot}
\title{Plot the SPC chart}
\usage{
rfl_spc_plot(data)
}
\arguments{
\item{data}{data from rfl_get_spc_model_output}

\item{pointSize}{Size fo the alerts and geom_points}

\item{...}{Variables sent to labs() for example title = "" and subtitle = "".}
}
\value{
a ggplot SPC chart
}
\description{
Create SPC chart ggplot
}
