% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{savedTemplateFile}
\alias{savedTemplateFile}
\title{savedTemplateFile: RFL SPC Reporting Template Workbook}
\format{
a workbook object from openxlsx
}
\usage{
savedTemplateFile
}
\description{
a dataset containing the ibject returned from  openxlsx::loadWorkbook('SPC Reporting Template.xlsx')
}
\keyword{datasets}
